W193: six MATCHes and a very deep push -- eight functions dropped below nd45, five below nd15.
Reconstruction from the decoded retail window continues to be the only thing
that moves large functions, and this wave it also produced the largest single
collapses of the campaign.

MATCHes:
  datCalendar 00187ec0  nd700 -> 0, frame -0x50 -> -0x60. The genuinely
        uninitialized parent argument recovered for every direct sprite call,
        and every depth argument corrected from 72.0f to retail's 50.0f. That
        restored the missing saved s4 and the frame slot.
  mt_evtMenu 00380200   nd173 -> 0. Both dispatcher callbacks recovered,
        absolute button-register loads, a typed selection-state subobject,
        retail's decrement-before-increment body layout, and the final
        confirm/cancel predicate orientation.
  itfMesManager 003a4dd0 nd173 -> 0. The absolute 0x34-stride entry, its typed
        +0x20 work subobject, direct cleanup calls, and a HIDDEN active-message
        argument that keeps the checked value in a0.
  k_clump 001a6c00      nd191 -> 0, frame -0xc0 -> -0xa0, window cleared. The
        real 0x44-byte addressed user-data callback context turned out to be a
        0x40-byte copied name followed by the count -- retail uses strcpy plus a
        count store where our source had invented a 0x54-byte memset aggregate.
  y_misc 00424730       nd205 -> 0. The real control flow recovered: validate
        type 1..11, query the count, return early when the index is already in
        range, otherwise scan 20 entries and update the result only when the
        entry is positive and the index equals the running count.
  sfl_result 001fc720   nd5 -> 0. The comparison operand spelling that keeps the
        quotient in v0 and selects retail's $at/$v0 threshold temporaries.

Functions that collapsed but did not close, with what the reconstruction
established:
  mdlFile 00323c20    nd190 -> **11**. The allocator loop counter is genuinely
        u16; a natural u16 increment with a signed promoted comparison restores
        retail's narrowing, address calculation, increment and branch sequence.
  h_sfdply 0010c1a0   nd260 -> **25**, and 456/512 -> exactly 512/512. Linked-tail
        ownership through the queue sentinel, the 256-entry 0x1dc pool with a
        4-byte ownership state plus 0x1d8 payload, signed-16 pool indexing, the
        exact field initialization order, back-link insertion, and interrupt-state
        restoration polarity.
  mt_evtMenu 00375900 nd213 -> **1**. Both real dispatcher callbacks, absolute
        button-register loads, and a typed menu-range subobject holding state,
        value and both bounds. The sole residual is one commutative `addu`.
  sfl_result 001fc870 nd171 -> **12**; k_encount 001d7c60 nd141 -> **28**;
  mt_sceneFunc 003bd8b0 nd171 -> **43**; itfMesManager 003a7dd0 nd126 -> **33**;
  mdlFile 00323e10 nd241 -> 133 (retail calls 00323c20 through a u32-first ABI at
        that site, so a per-site alias removed a false narrowing);
  fclShopMisc 00400a90 nd180 -> 77 and 003fb530 nd272 -> 143;
  frTagTable 003aec20 nd147 -> 91; y_runtime 00429d40 nd145 -> 57 and
  0043c7f0 nd172 -> 125; h_maestro 001140d0 / 00114af0 -- a genuine twin pair --
        both nd441 -> **342** and both now exactly at window, from typing the
        RGBA channels as u8 so retail's mask-then-shift-then-andi extraction is
        exact; mt_evtMain 00368f20 nd504 -> 421 on a reverse-declared linear
        switch reproducing retail's 0x13,0x0a,0x1c,0x1d,0x0d test order;
  btlTarget 002c4a90 nd292 -> 206 with its 0x10-short frame repaired exactly, and
        002c48a0 nd239 -> 206.

The mdlFile VU triplet also gave up its last structural difference: moving both
loop increments into the `for` clause emits retail's index increment before the
entry-pointer increment at offsets 552/556, taking 00335180 and 003377f0 to nd48
and 00336630 to nd59. What remains there is enumerated to the offset and is all
scheduling and packed-byte rotation.

Two audits worth recording as closed:
  * **All CONCAT residue is gone from the Event files.** No CONCAT11, CONCAT22 or
    CONCAT31 call sites remain in any of the eight; CONCAT13 survives only as
    unused legacy macro definitions.
  * mt_evtMain 00369a20's full 9,200-byte window was decoded and every global
    reference audited: `DAT_0069e1a2` is the only indexed record table and its
    per-site `char` view already gives retail's 12-byte stride. There is no
    second byte-versus-word stride defect in that function.
  * mt_evtMenu 0037dfc0 nd7 is closed with the diff spelled out instruction by
    instruction: only the two independent label loads and the first store are
    scheduled differently, which is the independently-ordered-memory-access
    class. Not a mystery any more.

Checkpoint: 6497 / 8211 first-party MATCH (79.1%). 117 first-party functions now
sit at nd <= 30. Zero lost, zero new over-window addresses, one window cleared,
zero compile errors, zero stale markers.

+++ src/h_maestro.c
@@ -2163,10 +2163,10 @@ void func_001140d0(f32 depth,
     f32 recipZ;
     f32 z;
     void (**setState)(u32, u32);
-    u32 r;
-    u32 g;
-    u32 b;
-    u32 a;
+    u8 r;
+    u8 g;
+    u8 b;
+    u8 a;
     s32 i;
 
     recipZ = 1.0f / kwlnGetMainCamera()->nearPlane;
@@ -2455,10 +2455,10 @@ void func_00114af0(f32 depth,
     f32 recipZ;
     f32 z;
     void (**setState)(u32, u32);
-    u32 r;
-    u32 g;
-    u32 b;
-    u32 a;
+    u8 r;
+    u8 g;
+    u8 b;
+    u8 a;
     s32 i;
 
     recipZ = 1.0f / kwlnGetMainCamera()->nearPlane;
